Server/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-BasicAuth==0.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
SQLAlchemy==1.4.18
Werkzeug==2.0.1
